Avoid writing to html when bikeshed fails
When using the bikeshed API, we should avoid overwriting the HTML when
there is a failure.  For example, malformed input will result in an
HTTP 400 response with no document body, leading to an empty
index.html when curl's output is redirected to index.html.  To avoid
this, we can have curl write the output file directly (with -o)
instead of using a shell redirect.  This way, if there is a failure,
index.html is not overwritten at all, and make will not see the HTML
as "up to date".